+++ src/jjjexperiment/inputs/heating.py
@@ -1,7 +1,7 @@
 # SeasonalLoadクラスは削除 - 新アーキテクチャではAcSetting + Entityパターンを使用
 # HeatingAcSetting + HeatQuantityを使用してください
 from .ac_setting import HeatingAcSetting
-from .ac_quantity_entity import HeatQuantity
+from .ac_quantity_service import HeatQuantity
 
 # 後方互換性のためのエイリアス
 SeasonalLoad = HeatQuantity
